refactor: Mover formatDate y formatTime a utils. #306
