+++ src/shared/hooks/games/preguntados/usePreguntadosGameActions.ts
@@ -1,4 +1,5 @@
 import { useMemo } from "react";
+import { formatTime } from "@/shared/utils/format";
 
 export const usePreguntadosGameActions = () => {
   const itemVariants = useMemo(
@@ -6,21 +7,11 @@ export const usePreguntadosGameActions = () => {
       hidden: { opacity: 0, y: 20 },
       visible: { opacity: 1, y: 0 },
     }),
-    []
+    [],
   );
 
   const getOptionLetter = (index: number) => {
     return String.fromCharCode(65 + index);
-  };
-
-  // TODO: Pasar a función utils para reusabilidad
-  const formatTime = (seconds: number) => {
-    if (seconds < 60) {
-      return `${seconds}s`;
-    }
-    const minutes = Math.floor(seconds / 60);
-    const remainingSeconds = seconds % 60;
-    return `${minutes}:${remainingSeconds.toString().padStart(2, "0")}`;
   };
 
   const getResultMessage = (isPassed: boolean) => {
