+++ src/student/components/DetailsGame/DetailsAhoracado/DetailsAhorcado.tsx
@@ -8,14 +8,13 @@ import {
   FaBullseye,
   FaCheck,
   FaHeart,
-  // FaRegChartBar,
   FaRegLightbulb,
   FaSync,
   FaTimes,
 } from "react-icons/fa";
 import { RiTextSnippet } from "react-icons/ri";
 import { GiBookshelf, GiBrain } from "react-icons/gi";
-// import { getPerformanceLevel } from "../../../utils/performance";
+
 export default function ActivityDetails() {
   const {
     gameConfig,
@@ -28,14 +27,6 @@ export default function ActivityDetails() {
     livesRemaining,
   } = useAhorcadoGame();
   const { currentActivity } = useActivityStudent();
-  // const formatTime = (seconds: number) => {
-  //   const mins = Math.floor(seconds / 60);
-  //   const secs = seconds % 60;
-  //   return `${mins}:${secs.toString().padStart(2, "0")}`;
-  // };
-
-  // const errorRate = wrongGuesses / maxErrors;
-  // const performance = getPerformanceLevel((1 - errorRate) * 100);
 
   const containerVariants: Variants = {
     hidden: { opacity: 0, scale: 0.9 },
@@ -107,7 +98,7 @@ export default function ActivityDetails() {
             <span className={styles.statValue}>
               {
                 guessedLetters.filter((letter) =>
-                  gameConfig?.word.toLowerCase().includes(letter.toLowerCase())
+                  gameConfig?.word.toLowerCase().includes(letter.toLowerCase()),
                 ).length
               }
             </span>
@@ -136,14 +127,6 @@ export default function ActivityDetails() {
           </div>
         </div>
 
-        {/* <div className={styles.statCard}>
-          <div className={styles.statIcon}>⏱️</div>
-          <div className={styles.statContent}>
-            <span className={styles.statLabel}>Tiempo Total</span>
-            <span className={styles.statValue}>{formatTime(timeSpent)}</span>
-          </div>
-        </div> */}
-
         <div className={styles.statCard}>
           <div className={styles.statIcon}>
             <FaSync />
@@ -155,18 +138,6 @@ export default function ActivityDetails() {
             </span>
           </div>
         </div>
-
-        {/* <div className={`${styles.statCard} ${styles.performanceCard}`}>
-          <div className={styles.statIcon}>{performance.icon}</div>
-          <div className={styles.statContent}>
-            <span className={styles.statLabel}>Rendimiento</span>
-            <span
-              className={`${styles.statValue} ${styles[performance.color]}`}
-            >
-              {performance.level} %
-            </span>
-          </div>
-        </div> */}
       </motion.div>
 
       {/* Letras Utilizadas */}
@@ -202,45 +173,6 @@ export default function ActivityDetails() {
           <p className={styles.noLetters}>No se utilizaron letras</p>
         )}
       </motion.div>
-
-      {/* Análisis de Estrategia */}
-      {/* <motion.div className={styles.analysisSection} variants={itemVariants}>
-        <h4>
-          <FaRegChartBar /> Análisis de tu Estrategia
-        </h4>
-        <div className={styles.analysisGrid}>
-          <div className={styles.analysisItem}>
-            <span className={styles.analysisLabel}>Eficiencia:</span>
-            <div className={styles.progressBar}>
-              <motion.div
-                className={styles.progressFill}
-                initial={{ width: 0 }}
-                animate={{
-                  width: `${Math.max(
-                    0,
-                    100 - (wrongGuesses / maxErrors) * 100
-                  )}%`,
-                }}
-                transition={{ delay: 1, duration: 1.5, ease: "easeOut" }}
-              />
-            </div>
-          </div>
-
-          <div className={styles.analysisItem}>
-            <span className={styles.analysisLabel}>Velocidad:</span>
-            <div className={styles.progressBar}>
-              <motion.div
-                className={styles.progressFill}
-                initial={{ width: 0 }}
-                animate={{
-                  width: `${Math.max(0, 100 - (timeSpent / 300) * 100)}%`, // Asumiendo 5 min como tiempo base
-                }}
-                transition={{ delay: 1.2, duration: 1.5, ease: "easeOut" }}
-              />
-            </div>
-          </div>
-        </div>
-      </motion.div> */}
 
       {/* Consejos para Mejorar */}
       {!isGameWon && (
