+++ src/student/hooks/activities/activityDetails/useDetailsFormatters.ts
@@ -1,28 +1,6 @@
+import { formatDate, formatDuration } from "@/shared/utils/format";
+
 export const useDetailsFormatters = () => {
-  const formatDate = (dateString: string) => {
-    try {
-      return new Date(dateString).toLocaleDateString("es-ES", {
-        year: "numeric",
-        month: "long",
-        day: "numeric",
-        hour: "2-digit",
-        minute: "2-digit",
-      });
-    } catch {
-      return dateString;
-    }
-  };
-
-  const formatTime = (minutes: number) => {
-    if (minutes < 60) {
-      return `${minutes} minutos`;
-    }
-    const hours = Math.floor(minutes / 60);
-    const remainingMinutes = minutes % 60;
-    return remainingMinutes > 0
-      ? `${hours}h ${remainingMinutes}m`
-      : `${hours} hora${hours > 1 ? "s" : ""}`;
-  };
-
+  const formatTime = formatDuration;
   return { formatDate, formatTime };
 };
