+++ src/student/hooks/activities/useActivityTimer.ts
@@ -1,4 +1,5 @@
 import { useState, useEffect, useCallback, useRef } from "react";
+import { formatTime } from "@/shared/utils/format";
 
 interface UseActivityTimerProps {
   maxTimeInMinutes: number;
@@ -48,21 +49,6 @@ export const useActivityTimer = ({
       }
     };
   }, [isActive, timeLeftInSeconds, onTimeUp]);
-
-  const formatTime = useCallback((seconds: number): string => {
-    const hours = Math.floor(seconds / 3600);
-    const minutes = Math.floor((seconds % 3600) / 60);
-    const secs = seconds % 60;
-
-    if (hours > 0) {
-      return `${hours.toString().padStart(2, "0")}:${minutes
-        .toString()
-        .padStart(2, "0")}:${secs.toString().padStart(2, "0")}`;
-    }
-    return `${minutes.toString().padStart(2, "0")}:${secs
-      .toString()
-      .padStart(2, "0")}`;
-  }, []);
 
   const resetTimer = useCallback(() => {
     setTimeLeftInSeconds(maxTimeInMinutes * 60);
